Extended to support the file upload widget.
The current form is remembered on the ctx as an iforms.IForm.

The file up load widget will render image urls that it knows how to process,
so requests to a __widget_res__ segment are forwarded on to the widget for it 
to process.

To prevent the creation of lots of temporary files a resource manager is 
created on the form. This is used by the file upload widget to hold onto 
uploaded files until the form has passed the widget validation. Once the
widget validation passes successfully the uploaded files are converted to
application format and the temporary files are deleted.

Conversion of the temp files to application format is controlled by a 
convertible protocol factory registered with the widget and exposed as a
'convertibleFactory' attribute on the widget.