Added the FileUploadWidget.
A widget that handles the uploading of image files. A convertible is registered
with the widget to handle the conversion from and to application format. The
widget uses the resource manager to manage the uploaded files until the form
has sucessfully validated.

The original form.data may contain an 'current' image. The file upload widget 
will render an image url that it expects to process. If no image has been 
uploaded and there is a 'current' image the widget will use the registered 
convertible to create a nevow static.Data object that contains the original
image. If there is an uploaded image it will ask the resource manager to return
a path so that a nevow static.File instance can be created.

Once the form has successfully validated, the form framework will use the
registered convertible factory to convert the temp uploaded file into 
application format, and the temp files are deleted.