OK, so the ResourceMixin was completely broken. The idea of a mixin that called super() was always going to be wrong! Also, the requirement to call the mixin's __init__() to get self registered as the form factory causes havoc with inheritance.
I'm starting to experiment with the idea of some sort of pluggable behaviour
that can be associated with a Page, although it's not complete in any way.

This patch adds the resource component that must be delegated to for IResource
methods so it can a) render and process the form, b) locate form resources.

It also fixes up the ResourceMixin to work in /nearly/ the same way as it did
but, due to the nature of the rend.Page class hierarchy and the mro of old-
style classes your page class must look something like this:

    class MyPage(forms.ResourceMixin, rend.Page):
        def __init__(self, *a, **k):
            rend.Page.__init__(self, *a, **k)
            forms.ResourceMixin.__init__(self, *a, **k)
            # More of your code

The __init__ is actually optional (the ResourceMixin's __init__ will super()
the call, but that's broken) too. The position of the mixin in the list of
super classes is important.