Merge -r 106:114 post-it-instead to branch.
This changes the way form posts are handled. In stead of using a magic child
resource it uses ... wait for it ... a POST method to the same resource.
Hurrah!

Unfortunately, the ResourceMixin is broken and always has been (sorry!). It
will be replaced sometime but it's working for how although the new version 
does change behaviour a little with certain inheritance hierarchies.