Add the concept of a Group of form items (Fields, other Groups, whatever). By default, they are rendered as a <fieldset>.
This means that form items may have a parent. The parent will always be a Group
instance at the moment but I may try to refactor the Form to make use of Group
later.

Make the form.data the result data too. It doesn't make much sense to
differentiate them and anyone comparing form.data with data in the form's
action callback is asking for trouble because form.data is nothing to do with
what the user saw.

Make the Field responsible for processing the request args, filling in errors
etc. The Field is also responsible for knowing what to do with immutable types.

Fix an invalid use of ctx.key.